Optimised reconciler

+++ script.js
@@ -1,27 +1,53 @@
 function createDomElements(todos){
     var parentElement = document.getElementById("mainArea");
-    parentElement.innerHTML = "";
-        for(var i =0; i<todos.length; i++){
-        var childElement = document.createElement("div");
 
-        var grandChildElement1 = document.createElement("span");
-        grandChildElement1.innerHTML = todos[i].title;
+    var currentChildren = Array.from(parentElement.children);
+    console.log(currentChildren);
 
-        var grandChildElement2 = document.createElement("span");
-        grandChildElement2.innerHTML = todos[i].description;
+    let added = 0, deleted=0, updated=0;
 
-        var grandChildElement3 = document.createElement("button");
-        grandChildElement3.innerHTML = "Delete";
+    todos.forEach(function (item){
+        var existingChild = currentChildren.find(function (child){
+            return child.dataset.id === String(item.id);
+        })
 
-        grandChildElement3.setAttribute("onclick", "deleteTodo("+ todos[i].id + ")");
+        if(existingChild){
+            updated++;
 
-        childElement.appendChild(grandChildElement1);
-        childElement.appendChild(grandChildElement2);
-        childElement.appendChild(grandChildElement3);
+            existingChild.children[0].innerHTML = item.title;
+            existingChild.children[1].innerHTML = item.description;
 
-        parentElement.appendChild(childElement);
+            currentChildren = currentChildren.filter(function (child){
+                return child != existingChild;
+            });
 
-    }
+        } else {
+            added++;
+            var childElement = document.createElement("div");
+            childElement.dataset.id = item.id;
+
+            var grandChildElement1 = document.createElement("span");
+            grandChildElement1.innerHTML = item.title;
+
+            var grandChildElement2 = document.createElement("span");
+            grandChildElement2.innerHTML = item.description;
+
+            var grandChildElement3 = document.createElement("button");
+            grandChildElement3.innerHTML = "Delete";
+            grandChildElement3.setAttribute("onclick","deleteTodo("+ todos[i].id + ")")
+            
+            childElement.appendChild(grandChildElement1);
+            childElement.appendChild(grandChildElement2);
+            childElement.appendChild(grandChildElement3);
+            parentElement.appendChild(childElement);
+
+        }
+    })
+
+    currentChildren.forEach(function (child){
+        deleted++;
+        parentElement.removeChild(child);
+    } )
 };
 
 window.setInterval(()=>{
